Ensure acts_as_token_authenticatable only applies changes to desired models, not ActiveRecord::Base
